@@ -1,14 +1,14 @@
 module github.com/philpearl/powerline-go
 
-go 1.12
+go 1.24
 
 require (
-	github.com/StackExchange/wmi v0.0.0-20181212234831-e0a55b97c705 // indirect
-	github.com/go-ole/go-ole v1.2.4 // indirect
 	github.com/mattn/go-runewidth v0.0.4
-	github.com/shirou/gopsutil v2.18.12+incompatible
 	golang.org/x/crypto v0.0.0-20190313024323-a1f597ede03a
 	golang.org/x/sys v0.0.0-20190318195719-6c81ef8f67ca
 	golang.org/x/text v0.3.0
+	gopkg.in/ini.v1 v1.67.0
 	gopkg.in/yaml.v2 v2.2.2
 )
+
+require github.com/stretchr/testify v1.10.0 // indirect
